LayoutTests/pointer-lock/pointerlock-interface-disabled.html

<!DOCTYPE html><!-- webkit-test-runner [ enablePointerLock=false ] -->
<head>
    <title>PointerLock interface tests.</title>
    <link rel="help" href="https://w3c.github.io/pointerlock/">
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<script>
test(function() {
    assert_false('exitPointerLock' in Document.prototype);
},'Document.prototype.exitPointerLock does not exist');

test(function() {
    assert_false('pointerLockElement' in Document.prototype);
},'Document.prototype.pointerLockElement does not exist');

test(function() {
    assert_false('onpointerlockchange' in Document.prototype);
},'Document.prototype.onpointerlockchange does not exist');

test(function() {
    assert_false('onpointerlockerror' in Document.prototype);
},'Document.prototype.onpointerlockerror does not exist');

test(function() {
    assert_false('requestPointerLock' in Element.prototype);
},'Element.prototype.requestPointerLock does not exist');
</script>
</body>
</html>
